Handle missing book and use absolute image path on delete

+++ controllers/book.js
@@ -1,5 +1,6 @@
 import Book from "../models/Book.js";
 import fs from "fs" // module natif de Node.js pour gérer des fichier
+import path from "node:path";
 
 export const createBook = async (req, res, next) => {
 
@@ -69,11 +70,14 @@ export const getBestRatingBooks = async (req, res, next) => {
 export const deleteBook = async (req, res, next) => {
     try {
         const book = await Book.findOne({ _id: req.params.id })
+        if(book === null){
+            return res.status(404).json({message : "Livre non trouvé"})
+        }
         if (book.userId != req.auth.userId) {
             res.status(401).json({ message: "Not authorized" })
         } else {
             const filename = book.imageUrl.split("/images/")[1]
-            fs.unlink(`./images/${filename}`, async () => {
+            fs.unlink(path.join(import.meta.dirname, "..", "images", filename) , async () => {
                 try {
                     const book = await Book.deleteOne({ _id: req.params.id })
                     res.status(200).json({ message: "Livre supprimé" })
